refactor(matching): 매칭 상세 파일 위치 정리
